Rename the wasi-http `struct` module to `types`
