+++ crates/wasi-http/src/types.rs
